refactor: 기술 용어 변경
