+++ tomcat/src/main/java/com/techcourse/servlet/RegisterServlet.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller;
+package com.techcourse.servlet;
 
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
